fix(pak): entry offsets are header-relative, not absolute — repairs every multi-entry pak
Neversoft PAK data offsets are relative to each entry's own header position
(Queen-Bee PakEditor: HeaderStart + FileOffset), with companion .pab data at
(resolved - pak length). The shipped absolute reads only worked for
single-entry sub-PAK tables (header position 0); every multi-entry pak
(qb.pak, cutscene mains, cas paks, worldzone paks) silently extracted
shifted, garbled payloads.

GC .apk.ngc: in-pak entries are header-relative too — the previous "hoisted
tiling" model was a numerically near-equivalent approximation, now replaced;
companion .mpk.ngc entries keep RAW stored offsets (first entry stores 0).

Signature-validated across all three THAW builds with
tools/diagnostics/pak_offset_check.py: PS2 12,120 + PC 12,756 + GC 14,325
payload hits, zero mismatches (absolute: qb.pak.ps2 scores 2/266 vs 266/266).

Fallout fixed here too: correctly-extracted zone tex blobs expose their real
16-byte header (u16 version 6 + counts), so the standard-format exclusion in
IsZoneTex now reads the full u32 — worldzone texture embedding actually
improves (z_bh: 967 textures vs the old partial set). Generated hex-offset
entry names now carry resolved offsets (z_ho test fixture renamed).

Note: Sample/Builds pak-extracted subtrees predate this fix and need
regeneration via tools/SampleGenerator.

+++ src/NeversoftMultitool/Core/Formats/Texture/Ps2Scene/ZoneTex/ThawZoneTexCoreDecoder.cs
@@ -333,8 +333,11 @@ internal static class ThawZoneTexCoreDecoder
     {
         if (data.Length < 0x200) return false;
 
-        // Must NOT be a standard format
-        var version = BitConverter.ToUInt16(data);
+        // Must NOT be a standard TEX/IMG format (u32 version 2-6 at offset 0).
+        // Correctly-extracted zone tex blobs carry a 16-byte header of their own
+        // (u16 version 6, u16 group count, u32 record count, u32 DMA chain offset,
+        // u32 size) — the nonzero count at +2 keeps the u32 read outside 2-6.
+        var version = BitConverter.ToUInt32(data);
         if (version is >= 2 and <= 6) return false;
 
         var (_, recordCount) = DiscoverRecordTable(data);
